requirements.txt
langchain==0.0.129
openai>=0.27.2
python-dotenv>=1.0.0
google-search-results>=2.4.2
colorama>=0.4.6
apify-client>=1.0.0
chromadb>=0.3.18
unstructured
faiss-cpu
gradio~=3.24.1
pypdf>=3.7.0
fastapi~=0.95.0
uvicorn~=0.21.1
pydantic~=1.10.7